requirements.txt
Pillow==9.2.0
colorama==0.4.5
timg==1.1.6
